requirements.txt
fastapi
revChatGPT
uvicorn
python-dotenv
